requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.2.0
tenacity>=8.2.3
streamlit>=1.30.0
